Refactor EmojiSelector
