+++ src/components/common/EmojiSelector/components/Emoji/index.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Text } from 'react-native';
+
+import Touchable from '../../../../animations/Touchable';
+import { charFromEmojiObject } from '../../utils';
+import styles from './styles';
+
+function Emoji({ onSelect, emojiObject, isLastEmoji }) {
+  const charEmoji = charFromEmojiObject(emojiObject);
+  const margin = { marginRight: isLastEmoji ? 0 : 10 };
+
+  return (
+    <Touchable onPress={() => onSelect(charEmoji)}>
+      <Text style={[styles.emojiText, margin]}>{charEmoji}</Text>
+    </Touchable>
+  );
+}
+
+export default Emoji;
