+++ src/components/common/EmojiSelector/components/Emoji/styles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+import { Colors } from '../../../../../constants/theme';
+
+export default StyleSheet.create({
+  emojiText: {
+    color: Colors.White.Pure,
+    fontSize: 30,
+  },
+});
