+++ src/components/common/EmojiSelector/components/EmojiBoard/styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  emojiBoard: {
+    height: 250,
+  },
+  emojiBoardRow: {
+    flex: 1,
+    flexDirection: 'row',
+    marginBottom: 6,
+    justifyContent: 'space-between',
+  },
+  lastBoardRow: {
+    justifyContent: null,
+  },
+});
