+++ src/components/common/EmojiSelector/components/EmojiCategoryTitle/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Text } from 'react-native';
+
+import styles from './styles';
+
+function EmojiCategoryTitle({ currentCategory }) {
+  return <Text style={styles.emojiCategoryTytle}>{currentCategory.name}</Text>;
+}
+
+export default EmojiCategoryTitle;
