+++ src/components/common/EmojiSelector/components/EmojiCategoryTitle/styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  emojiCategoryTytle: {
+    fontSize: 16,
+    marginBottom: 3,
+    color: 'rgba(235, 235, 245, 0.6)',
+  },
+});
