+++ src/components/common/EmojiSelector/styles.js
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  emojiSelector: {
+    marginTop: 24,
+    marginLeft: 7,
+    marginRight: 7,
+    marginBottom: 24,
+  },
+});
